requirements.txt
absl-py==2.1.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.48.1
fsspec==2024.2.0
idna==3.6
jax==0.4.25
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mediapipe==0.10.11
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.6
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
